+++ src/main/kotlin/icu/windea/pls/cwt/config/CwtKvConfig.kt
@@ -38,7 +38,7 @@ abstract class CwtKvConfig<out T : PsiElement> : CwtConfig<T> {
 	}
 	
 	
-	private fun inferScopeMap(): Map<String,String> {
+	private fun inferScopeMap(): MutableMap<String,String> {
 		//option的名字可能是：replace_scope/replace_scopes/push_scope
 		//对应的option可能位于：alias规则定义上，上一级definitionProperty规则定义上，definition规则定义上，subtype规则定义上
 		val result:MutableMap<String,String> = mutableMapOf()