+++ src/main/kotlin/icu/windea/pls/cwt/editor/CwtDocumentationProvider.kt
@@ -2,9 +2,12 @@ package icu.windea.pls.cwt.editor
 
 import com.intellij.lang.documentation.*
 import com.intellij.psi.*
+import com.intellij.psi.util.*
 import icu.windea.pls.*
 import icu.windea.pls.cwt.config.*
+import icu.windea.pls.cwt.expression.*
 import icu.windea.pls.cwt.psi.*
+import icu.windea.pls.script.psi.*
 import java.util.*
 
 class CwtDocumentationProvider : AbstractDocumentationProvider() {
@@ -24,13 +27,27 @@ class CwtDocumentationProvider : AbstractDocumentationProvider() {
 				if(configType != null) append("(").append(configType.text).append(") ")
 				else append ("(property) ")
 				append("<b>").append(name.escapeXmlOrAnonymous()).append("</b>")
-				//为localisation_command/modifier提供关于scope的额外文档注释
 				when(configType) {
+					//为definitionProperty提供特殊的文档注释（expression & scope)
+					null -> {
+						if(isDefinitionProperty(element, originalElement)){
+							val propertyElement = originalElement?.parent?.parent as? ParadoxScriptProperty ?:return@buildString
+							val config = propertyElement.propertyConfig ?:return@buildString
+							appendBr().append("(expression) ").append(config.keyExpression)
+								.append(" = ").append(config.valueExpression)
+							val scopeMap = mergeScope(config.scopeMap,propertyElement.definitionPropertyInfo?.scope)
+							for((sk, sv) in scopeMap) {
+								appendBr().append("(scope) ").append(sk).append(" = ").append(sv)
+							}
+						}
+					}
+					//为localisation_command提供关于scope的额外文档注释
 					CwtConfigType.LocalisationCommand -> {
 						//可以直接从psi文本中读取
 						val supportedScopesText = element.value?.text ?: "?"
 						appendBr().append("supported_scopes = $supportedScopesText")
 					}
+					//为modifier提供关于scope的额外文档注释
 					CwtConfigType.Modifier -> {
 						val gameType = originalElement?.gameType?: return@buildString
 						val configGroup = getConfig()[gameType] ?: return@buildString
@@ -73,13 +90,27 @@ class CwtDocumentationProvider : AbstractDocumentationProvider() {
 				if(configType != null) append("(").append(configType.text).append(") ")
 				else append ("(property) ")
 				append("<b>").append(name.escapeXmlOrAnonymous()).append("</b>")
-				//为localisation_command/modifier提供关于scope的额外文档注释
 				when(configType) {
+					//为definitionProperty提供特殊的文档注释（expression & scope)
+					null -> {
+						if(isDefinitionProperty(element, originalElement)){
+							val propertyElement = originalElement?.parent?.parent as? ParadoxScriptProperty ?:return@buildString
+							val config = propertyElement.propertyConfig ?:return@buildString
+							appendBr().append("(expression) ").append(config.keyExpression)
+								.append(" = ").append(config.valueExpression)
+							val scopeMap = mergeScope(config.scopeMap,propertyElement.definitionPropertyInfo?.scope)
+							for((sk, sv) in scopeMap) {
+								appendBr().append("(scope) ").append(sk).append(" = ").append(sv)
+							}
+						}
+					}
+					//为localisation_command提供关于scope的额外文档注释
 					CwtConfigType.LocalisationCommand -> {
 						//可以直接从psi文本中读取
 						val supportedScopesText = element.value?.text ?: "?"
 						appendBr().append("supported_scopes = $supportedScopesText")
 					}
+					//为modifier提供关于scope的额外文档注释
 					CwtConfigType.Modifier -> {
 						val gameType = originalElement?.gameType?: return@buildString
 						val configGroup = getConfig()[gameType] ?: return@buildString
@@ -138,6 +169,12 @@ class CwtDocumentationProvider : AbstractDocumentationProvider() {
 			}
 		}
 		return documentationLines?.joinToString("\n")
+	}
+	
+	private fun isDefinitionProperty(element:CwtProperty,originalElement: PsiElement?):Boolean{
+		//如果element.text对应的expression的类型是const，并且originalElement是scriptPropertyKeyId
+		return CwtKeyExpression.resolve(element.text.unquote()).type == CwtKeyExpression.Type.Constant
+			&&  originalElement?.parent is ParadoxScriptPropertyKey
 	}
 	
 	override fun getDocumentationElementForLink(psiManager: PsiManager?, link: String?, context: PsiElement?): PsiElement? {